Rename to OcisPhpSdk to match repo name
