+++ composer.json
@@ -1,10 +1,10 @@
 {
-    "name": "owncloud/ocis-sdk-php",
+    "name": "owncloud/ocis-php-sdk",
     "type": "library",
     "license": "Apache 2.0",
     "autoload": {
         "psr-4": {
-            "Owncloud\\OcisSdkPhp\\": "src/"
+            "Owncloud\\OcisPhpSdk\\": "src/"
         }
     },
     "repositories": [
