+++ src/DriveOrder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Owncloud\OcisSdkPhp;
+namespace Owncloud\OcisPhpSdk;
 
 abstract class DriveOrder
 {
@@ -9,7 +9,7 @@ abstract class DriveOrder
 
     public static function isOrderValid(?string $order): bool
     {
-        $reflector = new \ReflectionClass('Owncloud\OcisSdkPhp\DriveOrder');
+        $reflector = new \ReflectionClass('Owncloud\OcisPhpSdk\DriveOrder');
         if (!in_array($order, array_merge([null], $reflector->getConstants()))) {
             return false;
         }
